+++ home-manager/default.nix
@@ -3,6 +3,7 @@
     ./modules/git.nix
     ./modules/gpg.nix
     ./modules/identity.nix
+    ./modules/pms.nix
     ./modules/zsh.nix
   ];
 }